fix: 修复从数据行解析 ANI_ID 时的类型转换问题

+++ src/main/java/Main/FetchTask/FetchAnimeInfoTask.java
@@ -64,7 +64,7 @@ public class FetchAnimeInfoTask extends FetchInfoTask {
             if(ani_id_index != -1) {
                 var ani_id_str = row[ani_id_index];
                 if(ani_id_str != null && !ani_id_str.isBlank()) {
-                    try { ani_id = Integer.parseInt(ani_id_str.trim()); }
+                    try { ani_id = (int) Double.parseDouble(ani_id_str.trim()); }
                     catch(NumberFormatException _) {}
                 }
             }
